:bug: There was a circular dependency!
This is a massive bug that doesn't flag.
