+++ source/workers/readWav.worker.js
@@ -4,7 +4,7 @@
 import registerPromiseWorker from "promise-worker-transferable/register";
 
 // Helpers
-import { readArrayBufferPromise } from "../help/wav";
+import { readArrayBufferPromise } from "../help/fileRead";
 
 // Register Worker
 export default registerPromiseWorker(readArrayBufferPromise);
